+++ docs/changelog.md
@@ -8,5 +8,7 @@
 ### Voting Patterns App
 * Added sort legislators by district electorate lean
 
+### Legislator Lookup App
+* Created basic version which displays name and party of legislators based on address input
 
 ### ETL Pipeline